aspnetMVCExample.cshtml
This is an example using ASP.NET MVC using a ViewModel to use the gauges.  This is used to display the amount of work left to do and how much has been done.  This example needs the Model and the ViewModel coded as this is not a working example due to all the data needed from external soruces.